Add ability to switch mode (one-way)
If the switch connected to PA.0 is pressed,
the input will read `HIGH` and switch the app
mode to `SwitchMode` so the LED now follows
the user switch state.

This is a one way toggle - app will not be
able to switch back to the blinky mode unless
the board is reset.